requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
ecdsa==0.18.0
email-verify==0.1.10
et-xmlfile==1.1.0
fastapi==0.105.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.2
pyasn1==0.5.1
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-jose==3.3.0
pytz==2024.1
requests==2.31.0
rsa==4.9
secure-smtplib==0.1.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.1.0
uvicorn==0.24.0.post1
